alway go new onboarding swiper once
